Removed streamable_to_stationary
This functionality was removed since the default modules shall be stationary by the class definition of NeuralNetwork. Conversion from streamable to stationary is rarely needed in practice and would only require unneccessary extra code to make every new module class convertible. if it is then it can be done by saving the parameters of the streamable network and loading them into a fresh instance of the stationary network.